move amo items to src and move tests to root.
