+++ .c8rc
@@ -1,7 +1,7 @@
 {
   "all": true,
-  "include": ["out/**"],
-  "exclude": ["**/node_modules/**", "**/test/**", "out/amo/types.js", "out/amo/views/welcomeView.js", "**/config/**"],
+  "include": ["out/src/**"],
+  "exclude": ["**/node_modules/**", "**/test/**", "out/src/types.js", "out/src/views/welcomeView.js", "**/config/**"],
   "reporter": ["lcov", "text"],
   "lines": 5
 }
