+++ src/commands/updateTaskbar.ts
