+++ src/utils/addonCache.ts
