+++ src/utils/addonVersions.ts
@@ -3,7 +3,7 @@ import * as vscode from "vscode";
 
 import { showErrorMessage } from "./processErrors";
 import { makeAuthHeader } from "./requestAuth";
-import constants from "../../config/config";
+import constants from "../config/config";
 import { addonVersion, errorMessages } from "../types";
 
 export async function getAddonVersions(input: string, next?: string) {
