+++ src/views/sidebarView.ts
